вторая неделя

+++ src/main/java/ru/skypro/homework/dto/User.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class User {
 
-    private Long id;            // id пользователя
+    private Integer id;            // id пользователя
     private String email;       // логин пользователя
     private String firstName;   // имя пользователя
     private String lastName;    // фамилия пользователя
